Moved concern of Account creation to it's own static factory method. Updating references for package restructure. Updating exception message for AccountNotFoundException to be more specific to the issue.
