+++ src/main/java/com/groupthree/bankapp/controller/AccountController.java
@@ -1,7 +1,7 @@
 package com.groupthree.bankapp.controller;
 
-import com.groupthree.bankapp.entity.Account;
-import com.groupthree.bankapp.entity.AccountRequest;
+import com.groupthree.bankapp.model.entity.Account;
+import com.groupthree.bankapp.model.request.AccountRequest;
 import com.groupthree.bankapp.exception.AccountNotFoundException;
 import com.groupthree.bankapp.exception.AccountTypeNotSupportedException;
 import com.groupthree.bankapp.service.AccountService;
