+++ src/main/java/com/groupthree/bankapp/controller/TransactionController.java
@@ -1,8 +1,7 @@
 package com.groupthree.bankapp.controller;
 
-import com.groupthree.bankapp.entity.Account;
-import com.groupthree.bankapp.entity.RegularAccount;
-import com.groupthree.bankapp.entity.TransactionRequest;
+import com.groupthree.bankapp.model.entity.Account;
+import com.groupthree.bankapp.model.request.TransactionRequest;
 import com.groupthree.bankapp.exception.AccountNotFoundException;
 import com.groupthree.bankapp.exception.TransactionTypeNotSupportedException;
 import com.groupthree.bankapp.service.TransactionService;
@@ -14,11 +13,13 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 public class TransactionController {
+
     @Autowired
     TransactionService service;
 
     @PostMapping("/accounts/{accId}/transactions")
-    public ResponseEntity<Account> processTransaction(@PathVariable String accId, @RequestBody TransactionRequest transactionRequest) throws TransactionTypeNotSupportedException, AccountNotFoundException {
+    public ResponseEntity<Account> processTransaction(@PathVariable String accId, @RequestBody TransactionRequest transactionRequest)
+            throws TransactionTypeNotSupportedException, AccountNotFoundException {
         Account account = service.processTransaction(transactionRequest.getType(), accId, transactionRequest.getAmount());
         return new ResponseEntity<Account>(account, new HttpHeaders(), HttpStatus.OK);
     }
