+++ src/main/java/com/groupthree/bankapp/service/AccountService.java
@@ -1,9 +1,9 @@
 package com.groupthree.bankapp.service;
 
-import com.groupthree.bankapp.entity.Account;
-import com.groupthree.bankapp.entity.CheckingAccount;
-import com.groupthree.bankapp.entity.InterestAccount;
-import com.groupthree.bankapp.entity.RegularAccount;
+import com.groupthree.bankapp.model.entity.Account;
+import com.groupthree.bankapp.model.entity.CheckingAccount;
+import com.groupthree.bankapp.model.entity.InterestAccount;
+import com.groupthree.bankapp.model.entity.RegularAccount;
 import com.groupthree.bankapp.exception.AccountNotFoundException;
 import com.groupthree.bankapp.exception.AccountTypeNotSupportedException;
 import com.groupthree.bankapp.repository.AccountRepository;
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class AccountService {
@@ -21,26 +22,7 @@ public class AccountService {
     AccountNumberGeneratorService accountNumberGeneratorService;
 
     public Account createAccount(String name, String accType) throws AccountTypeNotSupportedException {
-        Account newAcc;
-        switch (accType) {
-            case "regular" : {
-                newAcc = new RegularAccount();
-                break;
-            }
-            case "checking" : {
-                newAcc = new CheckingAccount();
-                break;
-            }
-            case "interest" : {
-                newAcc = new InterestAccount();
-                break;
-            }
-            default: {
-                throw new AccountTypeNotSupportedException();
-            }
-        }
-        newAcc.setName(name);
-        newAcc.setAcctNumber(accountNumberGeneratorService.generate());
+        Account newAcc = Account.of(accType, name, accountNumberGeneratorService.generate());
         repository.save(newAcc);
         return newAcc;
     }
